requirements.txt
certifi==2018.10.15
chardet==3.0.4
idna==2.7
python-decouple==3.1
requests==2.20.1
six==1.11.0
tapioca-harvest==2.0
tapioca-wrapper==0.5.3
urllib3==1.24.1
